requirements.txt
langchain
langchain-mistralai
langchain-community
langchain-huggingface
langgraph
python-dotenv
faiss-cpu
python-magic
pdfplumber